feat: open modal
