+++ src/app/app.routing.module.ts
@@ -2,12 +2,10 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { UserListComponent } from './components/users/user-list/user-list.component';
-import { ModalComponent } from './components/modal/modal.component';
 import { NotFoundComponent } from './errors/not-found/not-found.component';
 
 const routes: Routes = [
   { path: 'user-list', component: UserListComponent },
-  // { path: 'transaction', component: ModalComponent },
   { path: '**', component: NotFoundComponent }
 ]
