--- src/app/components/modal/modal.component.html
@@ -1,1 +1,0 @@
-<p>modal works!</p>
