--- src/app/components/modal/modal.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-modal',
-  templateUrl: './modal.component.html',
-  styleUrls: ['./modal.component.scss']
-})
-export class ModalComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
