+++ src/app/components/users/user-list/user/user.component.html
@@ -11,7 +11,7 @@
     </div>
   </div>
 
-  <button type="button" class="pay-button">
-    Pagar
-  </button>
+  <button type="button" class="pay-button" (click)="showDialog = !showDialog">Pagar</button>
 </div>
+
+<app-modal [(visible)]="showDialog" title="My First Mpdal"></app-modal>
