+++ src/app/components/users/user-list/user/user.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { User } from '../../user';
 import { UserService } from '../../user.service';
+import { ModalService } from '../../modal.service';
+import { ModalComponent } from './modal/modal.component';
 
 @Component({
   selector: 'app-user',
@@ -9,8 +11,9 @@ import { UserService } from '../../user.service';
 })
 export class UserComponent implements OnInit {
   @Input() users: User[] = [];
+  // showDialog
 
-  constructor(private userService: UserService) { }
+  constructor(private userService: UserService, private modalService: ModalService) { }
 
   ngOnInit() {
     this.userService
@@ -18,4 +21,29 @@ export class UserComponent implements OnInit {
       .subscribe(users => this.users = users);
   }
 
+  // openModal() {
+  //   console.log('abriu modal');
+  //   this.showDialog = !this.showDialog
+  //   console.log('saiu modal');
+  // }
+
+  // openModal() {
+  //   this.modalService.open(ModalComponent);
+  // }
+
+  // openModal() {
+  //   const modalRef = this.modalService.open(
+  //     ModalComponent, 
+  //     { title: 'My dynamic title', 
+  //       // input valor transferência?
+  //       // select cartão?
+  //       message: 'My dynamic message' }
+  //   );
+
+  //   modalRef.onResult().subscribe();
+  // }
+
+  // openModal(id: string) {
+  //   this.modalService.open(id);
+  // }
 }
