+++ src/app/components/users/users.module.ts
@@ -5,12 +5,14 @@ import { HttpClientModule } from '@angular/common/http';
 import { UserListComponent } from './user-list/user-list.component';
 import { UserComponent } from './user-list/user/user.component';
 import { UserPhotoComponent } from './user-list/user-photo/user-photo.component';
+import { ModalComponent } from './user-list/user/modal/modal.component';
 
 @NgModule({
   declarations: [ 
     UserComponent, 
     UserListComponent,
-    UserPhotoComponent 
+    UserPhotoComponent,
+    ModalComponent 
   ],
   imports: [ 
     CommonModule,
